refactor: move 'define-store' to folder
