+++ src/index.ts
@@ -1,5 +1,5 @@
 export { setupFeathersPinia } from './setup'
-export { defineStore } from './define-store'
+export { defineStore } from './service-store/define-store'
 export { defineAuthStore } from './define-auth-store'
 export { makeServiceStore, BaseModel } from './service-store/index'
 export { makeState } from './service-store/make-state'
@@ -15,3 +15,6 @@ export { usePagination } from './use-pagination'
 export { syncWithStorage } from './storage-sync'
 export { syncWithStorageCompressed } from './storage-sync-compressed'
 export { clearStorage } from './clear-storage'
+
+export * from "./types";
+export * from "./service-store/types";
